fix/feat: 系统稳定性与功能体验综合优化及修复
【业务视角】
1. 修复了鼠标右键拖动手势操作时导致的鼠标卡死和右键菜单被吞弃的问题，保障基础交互流畅性。
2. 解决了设置界面的白屏问题（修正了 Vite 资源相对路径），确保用户可以正常访问并修改应用配置。
3. 修复了多屏环境下的手势轨迹起点异常结团和多屏坐标偏移问题。
4. 修复并优化了托盘消息提醒、应用程序图标展示等辅助体验问题。

【技术视角】
1. 修复了日志系统中 TraceId 失效的缺陷，并完善了手势操作的全链路结构化日志，提升了排查与定位能力。
2. 优化了钩子函数与手势引擎的底层热路径逻辑，实施了去阻塞改造与进程名快速查找，大幅降低 CPU 开销与系统延迟。
3. 解决了 PinWindow::close() 中存在的 use-after-free 内存自析构安全隐患。
4. 完善了项目工程建设：消除了两处 W4 级别警告，修复了 CI 编译问题（WinUtils 头文件缺失与 Lua sol 类型问题），增加了 vcpkg 二进制缓存来提升 CI 构建韧性。
5. 在打包部署层面，确保了部署产物带齐所有必要 DLL 文件，并完善了单元测试以保证代码质量。

+++ src/core/ipc/MessageBridge.cpp
@@ -275,7 +275,7 @@ void MessageBridge::registerBuiltinHandlers() {
                     GetModuleFileNameW(nullptr, exePath, MAX_PATH);
                     RegSetValueExW(hKey, L"EasyTools", 0, REG_SZ,
                                    reinterpret_cast<const BYTE*>(exePath),
-                                   (wcslen(exePath) + 1) * sizeof(wchar_t));
+                                   static_cast<DWORD>((wcslen(exePath) + 1) * sizeof(wchar_t)));
                 } else {
                     RegDeleteValueW(hKey, L"EasyTools");
                 }
